Identify which folders are git repos, clean them up and organize them. READ WARNING IN DESCRIPTION

This function will output a dataframe that will state whether folders within a specified folder are git repos, and their status. By status I mean, what's the currently checked-out branch, whether there are any changes noticed between the local repo and the remote. CAREFUL!!!!!!!!!!!!!!!!!!!!!!!!!! This command will also run gc, delete local merged branches and prune your remote !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

gitOrganizer(paff = "..", matchingText = ".")

Arguments

paff
path to the file, Default: '..'

matchingText
PARAM_DESCRIPTION, Default: '.'

Value

will return a dataframe with the following fields: name,gitRepo (T/F), branch(currently checked out branch), status (ok, behind or ahead of the remote), toCommit (local changes).

Also, the console will print out folders that are NOT git repos, folders that ARE git repos but are NOT on the master branch, and git repos that are in general not synced with the remote.

Details

DETAILS

Examples

if (FALSE) {
if(interactive()){
 #EXAMPLE1
 }
}